Remove the constraint check
The constraint checking of "a correct feedback cannot be applied after a
wrong-path feedback" is discarded. The reason is that there might be
several control dominators for a step with wrong-path feedback, and
Microbat can only recommend one of them. Thus, the user may have the
freedom to provide correct feedback after wrong-path feedback.