fix a bug of null variable ID
In the past, given a reachable local variable var, I think this
reachable local variable must be located in this static defined scope.
However, it is not true for constructor. If a constructor has a
parameter p, p can be reached from a field declaration outside the
constructor. This case cause the bug of null variable ID.