fix a trace hierarchy problem
I am not sure the reason yet. When parsing the method entry/exit event,
I need to ignore the <clinit> method event, i.e., the static method of
initializing a class when invoking the static method of this class. With
my code, the entry/exit the <clinit> methods seems to be not paired.
Therefore, the hierarchy structure of method caller and callee can be
sometimes wrong if the <clinit> is not ignored.

The effort to find out why this happens is necessary. However, I am too
tired now. I will leave this problem in the future.