support alias variable ID for array element
namely, an array element variable can have two variable id, one for heap
address (if it is not of primitive type) and one for index (parent
id+index). The latter is used for trace the data dependency by setting
null, e.g., a[0]=null. The former variable id cannot trace such a case.