Improve the array variable name searching algorithm
Scan the previous instructions to look for which variable used before
stands for the array name. The basic idea is that, given an array
instruction AINS such as aaload or aastore, we find its nearest previous
instruction handling an array variable. The name of handled array
variable is assumed to be the name of the given array. However, we
should take care that the instructions popping an array out of frame
stack. If we detect k such instructions, the name of variable handled by
k-th instruction from AINS should be the name we need. We record k in a
local variable popArrayTime.